refactor(test): 拆分超限测试文件
**修复**:
- test_handoff_recorder_unified.py: 修复行过长（88字符限制）
- test_handoff_command.py: 347 行 → 290 行
  - 提取 TestCheckCommand 到 test_check_command.py
- test_review_runner.py: 321 行 → 296 行
  - 提取 TestAgentResult 到 test_agent_result.py

**新增文件**:
- test_check_command.py (68 行): Check 命令测试
- test_agent_result.py (31 行): AgentResult 数据类测试

**成果**:
- ✅ 所有测试文件 ≤ 300 行
- ✅ 符合 LOC 质量标准
- ✅ 提高测试可维护性
